-Fixed test

+++ Src/Newtonsoft.Json/Linq/JValue.cs
@@ -724,7 +724,8 @@ namespace Newtonsoft.Json.Linq
         {
             int valueHashCode = (_value != null) ? _value.GetHashCode() : 0;
 
-            return _valueType.GetHashCode() ^ valueHashCode;
+            // GetHashCode on an enum boxes so cast to int
+            return ((int)_valueType).GetHashCode() ^ valueHashCode;
         }
 
         private static bool ValuesEquals(JValue v1, JValue v2)
